requirements.txt
aiohttp>=3.13.3
bilibili-api-python>=17.4.1
browser-cookie3>=0.20.1
